changed types for type safety

+++ include/point.h
@@ -6,13 +6,15 @@
 #ifndef POINT_H
 #define POINT_H
 
+#include <stdlib.h>
+
 /**
  * \struct point_t
  * \brief Défini un point dans un espace 2D
  */
 typedef struct {
-	int x; /*!< Coordonnées X du point */
-	int y; /*!< Coordonnées Y du point */
+	size_t x; /*!< Coordonnées X du point */
+	size_t y; /*!< Coordonnées Y du point */
 } point_t;
 
 #endif /* POINT_H */
